app/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-jet-reboot==1.3.3
psycopg2-binary==2.9.6

